+++ Assets/Prefabs/green-toon-race-car.prefab
@@ -613,8 +613,8 @@ Rigidbody:
   m_GameObject: {fileID: 161410}
   serializedVersion: 2
   m_Mass: 100
-  m_Drag: 0
-  m_AngularDrag: 0.05
+  m_Drag: 0.05
+  m_AngularDrag: 0
   m_UseGravity: 1
   m_IsKinematic: 0
   m_Interpolate: 0
@@ -687,6 +687,17 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ced43a2d1cab27a4ca99af999b9b125a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  frontLeftCollider: {fileID: 0}
+  frontRightCollider: {fileID: 0}
+  backLeftCollider: {fileID: 0}
+  backRightCollider: {fileID: 0}
+  frontLeftWheel: {fileID: 0}
+  frontRightWheel: {fileID: 0}
+  backLeftWheel: {fileID: 0}
+  backRightWheel: {fileID: 0}
+  speed: 1000
+  braking: 5000
+  turning: 45
 --- !u!114 &11438046
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -772,7 +783,19 @@ Prefab:
   serializedVersion: 2
   m_Modification:
     m_TransformParent: {fileID: 0}
-    m_Modifications: []
+    m_Modifications:
+    - target: {fileID: 0}
+      propertyPath: m_Mass
+      value: 100
+      objectReference: {fileID: 0}
+    - target: {fileID: 0}
+      propertyPath: m_Drag
+      value: 0.05
+      objectReference: {fileID: 0}
+    - target: {fileID: 0}
+      propertyPath: m_AngularDrag
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 0}
   m_RootGameObject: {fileID: 161410}
