+++ Assets/Scripts/CarMotor/CarMotor.cs
@@ -4,106 +4,81 @@ using System;
 
 public class CarMotor : MonoBehaviour, IMovementController
 {
+    private SimpleCarMotorController carController;
 
-    bool isFlipped = false;
-    public CarMotorController carController;
-    Rigidbody rb;
-    private float maxSpeed = 200f;
-    bool braking = true;
-    void Start()
-    {
-        rb = GetComponent<Rigidbody>();
-    }
+    //Wheel Colliders
+    public WheelCollider frontLeftCollider;
+    public WheelCollider frontRightCollider;
+    public WheelCollider backLeftCollider;
+    public WheelCollider backRightCollider;
+
+    //Wheel gameobjects
+    public GameObject frontLeftWheel;
+    public GameObject frontRightWheel;
+    public GameObject backLeftWheel;
+    public GameObject backRightWheel;
+
+    //Speed,braking,turning
+    public float speed = 100f;
+    public float braking = 100f;
+    public float turning = 30f;
+
     private void OnEnable()
     {
         // Set the Movement interface for the car controller to be this
-        carController = new CarMotorController();
+        carController = new SimpleCarMotorController();
         carController.SetMovementController(this);
     }
 
-
+    // Update is called once per frame
     void FixedUpdate()
     {
-        if (!isFlipped)
+        MoveCar();
+        MoveWheels();
+    }
+
+    public void MoveCar()
+    {
+        carController.ControlCar();
+    }
+
+    public void MoveWheels()
+    {
+        carController.ControlWheelRotation();
+    }
+
+    public void Drive()
+    {
+        //Moving the car
+        backLeftCollider.motorTorque = Input.GetAxis("Vertical") * speed;
+        backRightCollider.motorTorque = Input.GetAxis("Vertical") * speed;
+
+        //Set brakeTorque to zero to ensure not braking anymore
+        backLeftCollider.brakeTorque = 0;
+        backRightCollider.brakeTorque = 0;
+
+        //Turning the car
+        frontRightCollider.steerAngle = Input.GetAxis("Horizontal") * turning;
+        frontLeftCollider.steerAngle = Input.GetAxis("Horizontal") * turning;
+
+        //Braking
+        if (Input.GetKey(KeyCode.Space))
         {
-            //Keyboard inputs for controlling car
-            if (Input.GetKey(KeyCode.W) || Input.GetKey(KeyCode.UpArrow))
-                StartCoroutine(OnForward());
-            if (Input.GetKey(KeyCode.S) || Input.GetKey(KeyCode.DownArrow))
-                StartCoroutine(OnDown());
-            if (Input.GetKey(KeyCode.D) || Input.GetKey(KeyCode.RightArrow))
-                StartCoroutine(OnRight());
-            if (Input.GetKey(KeyCode.A) || Input.GetKey(KeyCode.LeftArrow))
-                StartCoroutine(OnLeft());
+            backLeftCollider.brakeTorque = braking;
+            backRightCollider.brakeTorque = braking;
         }
-        if (Vector3.Dot(transform.up, Vector3.down) > 0)
-            isFlipped = true;
-        if (Input.GetKey(KeyCode.R) && isFlipped)
-            StartCoroutine(OnFlipped());
     }
 
-    #region Enumerators
-
-    // These enumerators call functions on the CarMotorController class and all logic is handled there
-
-    public IEnumerator OnForward()
+    public void RotateWheels()
     {
-        carController.MoveForward();
-        yield return null;
-    }
+        //Turn wheels
+        frontLeftWheel.transform.localEulerAngles = new Vector3(frontLeftWheel.transform.localEulerAngles.x, -(180 - (frontLeftCollider.steerAngle - frontLeftWheel.transform.localEulerAngles.z)), frontLeftWheel.transform.localEulerAngles.z);
+        frontRightWheel.transform.localEulerAngles = new Vector3(frontRightWheel.transform.localEulerAngles.x, frontRightCollider.steerAngle - frontRightWheel.transform.localEulerAngles.z, frontRightWheel.transform.localEulerAngles.z);
 
-    public IEnumerator OnRight()
-    {
-        carController.TurnRight();
-        yield return null;
+        //Move wheels 
+        frontLeftWheel.transform.Rotate(frontLeftCollider.rpm / 60 * 360 * Time.deltaTime, 0, 0);
+        frontRightWheel.transform.Rotate(frontRightCollider.rpm / 60 * 360 * Time.deltaTime, 0, 0);
+        backLeftWheel.transform.Rotate(backLeftCollider.rpm / 60 * 360 * Time.deltaTime, 0, 0);
+        backRightWheel.transform.Rotate(backRightCollider.rpm / 60 * 360 * Time.deltaTime, 0, 0);
     }
-    public IEnumerator OnLeft()
-    {
-        carController.TurnLeft();
-        yield return null;
-    }
-    public IEnumerator OnDown()
-    {
-        carController.MoveBackward();
-        yield return null;
-    }
-
-    public IEnumerator OnFlipped()
-    {
-        Flip();
-        isFlipped = false;
-        yield return null;
-    }
-    #endregion
-
-    #region Movement Implementation
-
-    // This is where the car is actually moved or rotated
-
-    public void Translate(float value)
-    {
-        if (rb.velocity.sqrMagnitude > maxSpeed)
-            rb.AddForce(-value * transform.forward * 30, ForceMode.Force);
-
-        if (value < 0 && transform.InverseTransformDirection(rb.velocity).z > 0)
-            rb.AddForce(value * transform.forward, ForceMode.Force);
-        else
-            rb.AddForce(value * transform.forward * 30, ForceMode.Force);
-    }
-
-    public void Rotate(float value)
-    {
-        // Rotate around y axis
-        Vector3 v = Vector3.up * value * Time.deltaTime;
-        if (transform.InverseTransformDirection(rb.velocity).z != 0)
-            transform.Rotate(v);
-    }
-
-    public void Flip()
-    {
-        Quaternion rot = transform.rotation;
-        Vector3 v = new Vector3(rot.x, rot.y, 0);
-        transform.localEulerAngles = v;
-    }
-    #endregion
 }
