+++ Assets/Scripts/CarMotor/Interfaces/ICarMovementController.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public interface ICarMovementController
+{
+    
+}
