+++ Assets/Scripts/CarMotor/Interfaces/ICarMovementController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 76995b08b32bf67459c961fc689f4bd2
-timeCreated: 1456098263
+guid: 82188bf204518ee4b98ae73e02e763b7
+timeCreated: 1458862952
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
